+++ eugene/src/SensorPlugins/NG2/Sensor.NG2.h
@@ -1,3 +1,15 @@
+/*****************************************************************************/
+/*             Copyright (c) 2002 by INRA. All rights reserved.              */
+/*                 Redistribution is not permitted without                   */
+/*                 the express written permission of INRA.                   */
+/*                     Mail : tschiex@toulouse.inra.fr                       */
+/*---------------------------------------------------------------------------*/
+/* File         : EuGeneTk/SensorPlugins/NG2/Sensor.NG2.h                    */
+/* Description  : Sensor NetGene2                                            */
+/* Authors      : P.Bardou, S.Foissac, M.J.Cros, A.Moisan, T.Schiex          */
+/* History      : May 2003                                                   */
+/*****************************************************************************/
+
 #ifndef  SENSOR_NG2_H_INCLUDED
 #define  SENSOR_NG2_H_INCLUDED
 
@@ -11,8 +23,8 @@ class SensorNG2 : public Sensor
  private:
   int PositionGiveInfo;
 
-  std::vector<int>  vPosAccF, vPosAccR, vPosDonF, vPosDonR;
-  std::vector<REAL> vValAccF, vValAccR, vValDonF, vValDonR;
+  std::vector<int>    vPosAccF, vPosAccR, vPosDonF, vPosDonR;
+  std::vector<double> vValAccF, vValAccR, vValDonF, vValDonR;
 
   int iAccF, iAccR, iDonF, iDonR;
   double accB, accP, donB, donP;
